backend/requirements.txt
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
googlemaps==4.10.0
Pillow==10.1.0
psycopg2==2.9.9
PyJWT==2.8.0




       